refactor: navigation Route object 네이밍 수정
